requirements.txt
# Core dependencies
openai>=1.0.0
openai-whisper>=20240930
numpy>=1.20.0
torch>=2.0.0
tqdm>=4.64.0
python-dotenv>=1.0.0

# Optional dependencies
numba>=0.56.0  # Speeds up Whisper processing 